Get-InfoBlox - $SearchValue parameter was not being expanded in Begin block when referenced inside of an external script - but worked from command line.  Changed from $SearchValue to $PSBoundParameters["SearchValue"] and it works both ways now.
New-InfoBloxRR - added functionality for adding a host/A record as the next available IP in a range.  Noticed that on networks > /24 (/23, etc.), the Network get next available function won't go beyond the first /24.  Creating Ranges within the network for each /24 allows you to reference these for next available IP.  Need to use long form of the function to reference the range, because I need to skip .0 addresses in the range.

Added some helper scripts for getting AD DNS RRs/Zones via WMI for cross-domain pulling, Test-ConnectionAsync from MS PS Gallery, and Compare-DNSZoneToIB, this will compare the WMI DNS, or the MS DNSServer module output to a corresponding zone/record type in InfoBlox to see if records are missing or incorrect - for migrations.