Sync changes from in-house rep
Starting managing this project in VSTS.  This is an update to get this public repo back in sync.

Added Get-InfoBloxFixedAddressReservation
Added Remove-InfoBloxFixedAddressReservation
Added Get-InfoBloxResourceRecord - this requires a _ref URI as opposed to searching based on filters like the ResourceRecordSet command.

Updates for TrustedCert policy - previously the scripts were relying on specific functions to be run first and setting up the TrustedCert policy - this is no longer the case, and now it runs in each function to check the policy and set it if necessary.

New-InfoBloxResourceRecord - now supports configuring DHCP.

Made some modifications to variable names in the DynamicParam blocks - some variable names were conflicting with the parameter names, which was forcing me to use $PSBoundParameters["VarName"].  There will be another update to switch these references over to just using the now not-conflicting parameter variables.

Other various bug fixes, such as sending credentials to functions when it is required, for when sessions are not used, and having the appropriate Mandatory attributes for parameters.