Moved helper functions into private folder
Updated New-InfoBloxResourceRecord to add functionality for excluding an IP Address/Range.  So for instance, if I have a network defined as 192.168.1.0/23, and I want to insert a new record, but I want to entirely skip the first /24 in this network, I can use the -Exclude parameter, and pass it 192.168.1.0/24 or 192.168.1.0-192.168.2.0, etc. to force it into the second network (the second example ensures it skips the .0 broadcast address)

Also added some coments to the end of code blocks that were not previously marked with a comment.